+++ R/zzz.R
@@ -42,6 +42,15 @@ rdhs_setup <- function() {
 
 }
 
+.onAttach <- function(libname, pkgname) {
+
+  str <- "Thank you for using rdhs. If you are using rdhs regularly
+  or for automated tasks, please register for your own API key by
+  emailing api@dhsprogram.com. \n
+  More info at {.url https://api.dhsprogram.com/#/introdevelop.html}"
+  packageStartupMessage(cli::cli_text(str))
+}
+
 # ask for user permission to write to Renviron
 #' @noRd
 ask_user_permission <- function(){